fixed Tripouille bug

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: igomes-h <italogholanda@gmail.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/13 01:53:34 by coder             #+#    #+#             */
-/*   Updated: 2021/11/20 22:49:44 by igomes-h         ###   ########.fr       */
+/*   Updated: 2021/11/21 17:26:48 by igomes-h         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,9 +25,9 @@ static int	print_by_type(va_list *args, char key)
 	if (key == 'p')
 		return (ptr_printer(va_arg(*args, unsigned long)));
 	if (key == 'x')
-		return(lohex_printer(va_arg(*args, unsigned long)));
+		return(lohex_printer(va_arg(*args, unsigned int)));
 	if (key == 'X')
-		return(uphex_printer(va_arg(*args, unsigned long)));
+		return(uphex_printer(va_arg(*args, unsigned int)));
 	if (key == '%')
 		return(pct_printer());
 	return (0);
