+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: igomes-h <italogholanda@gmail.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/13 02:06:54 by coder             #+#    #+#             */
-/*   Updated: 2021/11/21 10:44:20 by igomes-h         ###   ########.fr       */
+/*   Updated: 2021/11/21 17:30:23 by igomes-h         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,10 +25,10 @@ int	nbr_printer(int n);
 int	unint_printer(unsigned int n);
 int	char_printer(char c);
 int	str_printer(char *str);
-int	new_base_printer(unsigned long nbr, const char *base);
+int	new_base_printer(unsigned long long nbr, const char *base);
 int	ptr_printer(unsigned long ptr);
-int	lohex_printer(unsigned long hex);
-int	uphex_printer(unsigned long hex);
+int	lohex_printer(unsigned int hex);
+int	uphex_printer(unsigned int hex);
 int	pct_printer(void);
 
 #endif
